+++ The Internet Archive Companion/CollectionDataHandlerAndHeaderView.m
@@ -16,8 +16,8 @@
 
 @property (nonatomic, strong) IAJsonDataService *service;
 @property (nonatomic, strong) NSMutableArray *searchDocuments;
-@property (assign) int numFound;
-@property (assign) int start;
+@property (assign) NSInteger numFound;
+@property (assign) NSInteger start;
 @property (assign) BOOL didTriggerLoadMore;
 
 @end
@@ -86,7 +86,7 @@
 
 }
 
-- (int) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return [searchDocuments count];
 }
 
@@ -174,7 +174,7 @@
         didTriggerLoadMore = YES;
         start = start + 50;
 
-        [service setLoadMoreStart:[NSString stringWithFormat:@"%i", start]];
+        [service setLoadMoreStart:[NSString stringWithFormat:@"%li", (long)start]];
         [service fetchData];
         [[NSNotificationCenter defaultCenter] postNotificationName:@"ShowLoadingIndicator" object:[NSNumber numberWithBool:YES]];
